@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -47,6 +47,7 @@ import java.nio.file.Paths;
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import jdk.test.lib.JDKToolFinder;
 import jdk.test.lib.compiler.CompilerUtils;
+import jdk.test.lib.process.OutputAnalyzer;
 import jdk.test.lib.process.ProcessTools;
 import jdk.test.lib.util.JarUtils;
 
@@ -81,14 +82,14 @@ public class TestDriver {
         CompilerUtils.compile(srcDir.resolve("src").resolve("test"), targetDir);
 
         // Run tests
-        String java = JDKToolFinder.getTestJDKTool("java");
         String cp = targetDir.toString() + File.pathSeparator + jarFile;
         String[] tests = new String[]{"TestBug4361044", "TestBug4523159"};
         for (String test : tests) {
-            ProcessTools.executeCommand(java, "-cp", cp, test)
-                        .outputTo(System.out)
-                        .errorTo(System.out)
-                        .shouldHaveExitValue(0);
+            ProcessBuilder pb = ProcessTools.createTestJavaProcessBuilder("-cp", cp, test);
+            new OutputAnalyzer(pb.start())
+                    .outputTo(System.out)
+                    .errorTo(System.out)
+                    .shouldHaveExitValue(0);
         }
     }
 }
